+++ ts/features/bonus/siciliaVola/store/reducers/__test__/availableDestinations.test.ts
