+++ ts/features/bonus/siciliaVola/store/reducers/availableDestinations.ts
